Refactored and Renamed NodeRepository Tests
Rename tied to scope change of class. It will now focus specifcally on query tests. This allows us to move some Arrange logic to the constructor as it can now be shared
